+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add vhost to access log
 * Add http/2 support using the `spdy` module
 * Add websocket support for http-only proxies
+* Add `Domain` model, where you can enable wildcard support
 
 ## 0.2.0 (2017-09-01)